feat: descriptions for all data sources, resources and attributes
Setting the `Description` field on all data sources, resources and
attributes. This field is meant to be used by dev tooling.

In the absence of the `MarkdownDescription` field, `Description` will
also be used to generate documentation for the Terraform registry. This
is good enough to get started, but I'll add `MarkdownDescription` for
data sources and resources in a following PR to provide more
comprehensive usage documention.