fix: read method for helm and TF correctly sets vars (#7)
The `Read` method for both the Helm and Terraform resources was
appending to the TF model's attribute, instead of overwriting. This lead
to the attribute list being doubled, `terraform refresh` seeing a state
change when there was none.

The updated tests still fail, but now it's because the variable names
and values are somehow being wrapped in extra quotes. Will follow up
with a separate PR for that.