fix: app update method reads from api
We don't return nested data from the API's update endpoints, and the
provider was just ignoring that data. Even in cases where that data
isn't supposed to change, we should read it to make sure nothing falls
through the cracks.

Added a GET request to the update method fetch the entire app object
after updating it.